adds siteIndex worker & Single uploads bucket feature
siteIndex moves the responsibility of indexing a site onto the server. Instead of being handled in the browser via CMS making individual requests to the server, which then makes requests against elastic search. Instead, a single reindex signal is pushed, and the server downloads the firebase data for a site, as well as the current search index data, and produces a series of elastic search indexing commands which are batched and pushed at once.

Single uploads bucket makes one bucket for all webhook uploads. Instead of individual CMS documents uploading files into their respective buckets. The server uploads into the environment defined UPLOADS_BUCKET. In order to assist in this transition, there is a migrate-uploads task in the /bin/ directory. This task is responsible for copying all webhook uploads from individual buckets into the single uploads bucket.

+ adds siteIndex library to function as a worker. responsible for synchronizing data in firebase with the elastic search representation.
+ adds siteIndexWorker to grunt task list
+ adds listUploads function to cloudStorage, for migrating assets from multiple buckets into the UPLOADS_BUCKET
+ adds cors rule for webhook-assets.risd.systems to properly serve font files. Useful for the CMS
+ adds a copy object function cloud storage to support file migration
+ adds the siteSearchReindex command the the command delegator to handle